+++ CHANGELOG.md
@@ -36,6 +36,7 @@ Please mark all changes in change log and use the issue from GitHub
 -   \#3448 RHNSWFlatTest failed
 -   \#3511 SearchTask::nq() should not return 0
 -   \#3514 Search failed with regex tag name
+-   \#3533 scheduler/selector needs to judge the index type
 
 ## Feature
 -   \#2319 Redo metadata to support MVCC